cmds/core/ps: make clear it is only for linux at present
/proc is no longer there on freebsd, and on Plan 9,
it is just too different to work.
