+++ cmds/core/ps/ps_linux_test.go
